+++ CHANGES.md
@@ -11,6 +11,13 @@
 
 ## develop
 
+- [ADD] kvconf:initialize/3 を追加し、環境変数による設定の上書きをサポート
+  - Options マップでプレフィックスを指定可能
+    - `env_prefix => <<"SPAM">>`
+  - `env_prefix` が指定されている場合は環境変数による設定項目の上書きが有効になる
+  - `env_prefix` が `<<"SPAM">>` の場合、設定項目 `abc_efg` は環境変数 `SPAM_ABC_EFG` で上書できる
+  - @voluntas
+
 ### misc
 
 - [CHANGE] GitHub Actions のコンテナイメージを OTP を 28.1 / OpenSSL 3.6.0 に上げる